backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.1.0
